requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
celery>=5.2
redis>=4.5
django-redis
python-dotenv
